### 2022-04-12 #4315 @editorialbot
[PRE REVIEW]: NAIF PDS4 Bundler: A Python package to generate SPICE PDS4 archives
**Submitting author:** @mcostasitja (<a href="http://orcid.org/0000-0003-0055-2959">Marc Costa Sitja</a>)
**Repository:** https://github.com/NASA-PDS/naif-pds4-bundler
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
