### 2022-04-13 f8b05e1
Binary kernel determination added.
