+++ src/pds/naif_pds4_bundler/classes/setup.py
@@ -5,6 +5,7 @@ import logging
 import os
 import re
 import shutil
+import sys
 from os.path import dirname
 from pathlib import Path
 from xml.etree import cElementTree as ET
@@ -183,7 +184,7 @@ class Setup(object):
             self.date_format = "maklabel"
 
         #
-        # Check End of Line format and set EoL length.
+        # Check text End of Line format and set EoL length.
         #
         # CRs are added to all text, XML, and other PDS meta-files
         # present in PDS4 archives as dictated by the standard.
@@ -209,6 +210,26 @@ class Setup(object):
         self.eol_mk = "\n"
         self.eol_mk_len = 1
         self.eol_pds3 = "\r\n"
+
+        #
+        # Check and determine endianness.
+        #
+        if not hasattr(self, "binary_endianness"):
+            if self.pds_version == '4':
+                self.kernel_endianness = "little"
+            else:
+                self.kernel_endianness = "big"
+        else:
+            if self.binary_endianness.lower() == "little" or \
+                    self.binary_endianness.lower() == "ltl-ieee":
+                self.kernel_endianness = "little"
+            elif self.binary_endianness.lower() == "big" or \
+                    self.binary_endianness.lower() == "big-ieee":
+                self.kernel_endianness = "big"
+            else:
+                error_message("binary_endianness configuration parameter value must be"
+                              " 'big', 'BIG-IEEE', 'little' or 'LTL-IEEE'. Case is not"
+                              " sensitive.")
 
         #
         # Fill missing fields.
@@ -276,6 +297,35 @@ class Setup(object):
                 "[0-9]{4}-[0-9]{2}-[0-9]{2}T" "[0-9]{2}:[0-9]{2}:[0-9]{2}Z"
             )
             format = "YYYY-MM-DDThh:mm:ssZ"
+
+        #
+        # Check binary kernels endianness.
+        #
+        if self.pds_version == '4':
+            if  self.kernel_endianness == "little":
+                logging.info("-- Binary SPICE kernels expected to have LTL-IEEE (little endian) binary format.")
+            else:
+                logging.info("-- Binary SPICE kernels expected to have BIG-IEEE (little endian) binary format.")
+                logging.warning("-- NAIF strongly recommends to use LTL-IEEE (little endian) for binary kernels in PDS4 archives ")
+                logging.warning("   and enforces it if the archive is hosted by NAIF.")
+        else:
+            if self.kernel_endianness == "little":
+                logging.info("-- Binary SPICE kernels expected to have LTL-IEEE (little endian) binary format.")
+
+                logging.warning(
+                    "-- NAIF strongly recommends to use BIG-IEEE (big endian) for binary kernels in PDS3 archives "
+                )
+                logging.warning( "   and enforces it if the archive is hosted by NAIF.")
+                if sys.byteorder == "big":
+                    logging.warning("   Your system is BIG-IEEE (big endian): PDS3 Labels cannot be attached to binary kernels.")
+                    logging.warning("   If binary kernels are present this will result in an error.")
+            else:
+                logging.info("-- Binary SPICE kernels expected to have BIG-IEEE (big endian) binary format.")
+                if sys.byteorder != "big":
+                    logging.warning(
+                        "   Your system is LTL-IEEE (little endian): PDS3 Labels cannot be attached to binary kernels."
+                    )
+                    logging.warning("   If binary kernels are present this will result in an error.")
 
         #
         # Check Bundle increment start and finish times. For the two accepted
@@ -916,6 +966,9 @@ class Setup(object):
         self.fks = fks
         self.sclks = sclks
         self.lsk = lsk
+
+    def information_model_setup(self):
+        """Setup and check PDS4 Information Model related things."""
 
     def clear_run(self, debug=False):
         """Clears the files generated by a previous run.
