### 2022-04-13 95ef284
Added binary kernel endianness tests.
