+++ tests/naif_pds4_bundler/unittests/test_endianness.py
@@ -1,0 +1,254 @@
+"""Unit tests for binary kernel endianness.
+
+Binary kernel endianness of binary kernels needs to be the one specified via
+configuration and must be compatible with the machine being used. These tests
+ensure that this is the case.
+"""
+import os
+import shutil
+import sys
+
+from pds.naif_pds4_bundler.__main__ import main
+from pds.naif_pds4_bundler.utils.files import string_in_file
+
+
+def post_setup(self):
+    """Setup Test.
+
+    This method will be executed before each test function.
+    """
+    dirs = ["kernels/fk",
+            "kernels/lsk",
+            "kernels/spk",
+            "kernels/mk"
+            ]
+    for dir in dirs:
+        os.mkdir(dir)
+
+    shutil.copy2("../data/kernels/lsk/naif0012.tls", "kernels/lsk/")
+    shutil.copy2("../data/kernels/mk/m2020_v01.tm", "kernels/mk/")
+    shutil.copy2("../data/kernels/mk/m2020_chronos_v01.tm", "kernels/mk/")
+    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp", "kernels/spk/")
+
+
+def test_pds4_big_endianness(self):
+    """Test BIG-IEEE basic.
+
+    The test has a different logic depending on whether if the host machine is
+    LTL-IEEE or BIG-IEEE.
+    """
+    post_setup(self)
+
+    config = "../config/mars2020.xml"
+
+    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
+                 "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp")
+
+    if sys.byteorder == 'little':
+        try:
+            main(config, silent=True, log=True, faucet="Bundle")
+        except BaseException:
+            line_check = "Binary SPICE kernels expected to have LTL-IEEE (little endian) binary format."
+            if not string_in_file("working/mars2020_release_temp.log", line_check, 1):
+                raise BaseException
+
+            line_check = "The binary kernel is not readable by your machine."
+            if not string_in_file("working/mars2020_release_temp.log", line_check, 3):
+                raise BaseException
+        else:
+            raise BaseException
+
+    else:
+        main(config, silent=True, log=True, clear="Bundle")
+
+
+def test_pds4_big_endianness_config(self):
+    """Test BIG-IEEE when indicated via configuration."""
+    post_setup(self)
+
+    config = "../config/mars2020.xml"
+    updated_config = 'working/mars2020.xml'
+    with open(config, "r") as c:
+        with open(updated_config, "w") as n:
+            for line in c:
+                if "<end_of_line>CRLF</end_of_line>" in line:
+                    n.write(line)
+                    n.write("<binary_endianness>BIG-IEEE</binary_endianness>\n")
+                else:
+                    n.write(line)
+
+    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
+                 "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp")
+
+    main(updated_config, silent=True, log=True, faucet="Bundle")
+
+    if sys.byteorder == "big":
+        lines_check = ["NAIF strongly recommends to use LTL-IEEE (little endian) for binary kernels in PDS4 archives",
+                       "Binary SPICE kernels expected to have BIG-IEEE (little endian) binary format"]
+        for line_check in lines_check:
+            if not string_in_file("working/mars2020_release_01.log", line_check, 1):
+                raise BaseException
+    else:
+        line_check = "The binary kernel is big endian; this endianness is not supported by your machine."
[…]
