### 2022-04-13 48e2381
Added Binary Kernels Endianness section.
