### 2022-04-14 803a852
Streamlined support to test NPB with Docker.
