+++ tests/naif_pds4_bundler/test_unittests.py
@@ -40,7 +40,7 @@ class TestUnitTests(TestCase):
         cls.faucet = "list"
         cls.verbose = False
         cls.log = True
-        cls.tmp_dir = tempfile.TemporaryDirectory(dir="/Users/mcosta")
+        cls.tmp_dir = tempfile.TemporaryDirectory()
 
         shutil.copytree(os.sep.join(cls.test_dir.split(os.sep)),
                         cls.tmp_dir.name + '/naif_pds4_bundler')
