+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [v1.0.2-dev](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.0.2-dev) (2022-04-08)
+## [v1.0.2](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.0.2) (2022-04-08)
 
-[Full Changelog](https://github.com/NASA-PDS/naif-pds4-bundler/compare/v1.0.1...v1.0.2-dev)
+[Full Changelog](https://github.com/NASA-PDS/naif-pds4-bundler/compare/v1.0.1...v1.0.2)
 
 ## [v1.0.1](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.0.1) (2022-04-01)
 
