### 2022-04-19 #4327 @SimonMolinsky
Hello @mbobra , Hi @athulpg007 ,

Thank you for joining me into the reviewers team 👋

When I've submitted paper to JOSS, authors were visible in the review thread (issue). Is this still the case? Should we interact with authors here, or should we create issues in the package repository?
