### 2022-04-19 7183ca5
Tests do not remove directory trees but move them.
