### 2022-04-21 bef42af
Refactored tearDown class to avoid deleting directories other than the test temporary ones.
