+++ tests/naif_pds4_bundler/test_unittests.py
@@ -47,7 +47,10 @@ class TestUnitTests(TestCase):
         shutil.copytree(os.sep.join(cls.test_dir.split(os.sep)[:-2]) +
                         "/src/pds/naif_pds4_bundler/templates/1.5.0.0",
                         cls.tmp_dir.name + '/naif_pds4_bundler/templates/1.5.0.0')
-        os.chdir(cls.tmp_dir.name + "/naif_pds4_bundler/unittests/")
+
+        tests_dir = cls.tmp_dir.name + "/naif_pds4_bundler/unittests/"
+        print(f"       Tests data on: {cls.tmp_dir.name}")
+        os.chdir(tests_dir)
 
     @classmethod
     def tearDownClass(cls):
@@ -84,10 +87,11 @@ class TestUnitTests(TestCase):
         This method will be executed after each test function.
         """
         unittest.TestCase.tearDown(self)
-        dirs = next(os.walk('.'))[1]
+        test_dir = self.tmp_dir.name + "/naif_pds4_bundler/unittests/"
+        dirs = next(os.walk(test_dir))[1]
         for dir in dirs:
             try:
-                shutil.rmtree(dir, ignore_errors=True)
+                shutil.rmtree(test_dir + dir)
             except BaseException:
                 pass
 
