### 2022-04-21 ec117ad
Added draft of product checks after kernel list generation.
