+++ src/pds/naif_pds4_bundler/utils/files.py
@@ -5,6 +5,7 @@ import glob
 import hashlib
 import json
 import logging
+import sys
 import os
 import re
 import shutil
@@ -906,3 +907,156 @@ def format_multiple_values(value):
         value = value[:-2] + '\n' + ' ' * 31 + "}\n"
 
     return value
+
+
+def product_mapping(name, setup):
+    """Obtain the kernel mapping.
+
+
+    :return: Kernel Mapping
+    :rtype: str
+    """
+    kernel_list_file = (
+            setup.working_directory
+            + os.sep
+            + f"{setup.mission_acronym}_{setup.run_type}_"
+              f"{int(setup.release):02d}.kernel_list"
+    )
+
+    get_map = False
+    mapping = False
+
+    with open(kernel_list_file, 'r') as lst:
+        for line in lst:
+            if name in line:
+                get_map = True
+            if get_map and "MAPPING" in line:
+                mapping = line.split("=")[-1].strip()
+                get_map = False
+
+    if not mapping:
+        error_message(
+            f"{name} does not have mapping on {kernel_list_file}.",
+            setup=setup,
+        )
+
+    return mapping
+
+
+def check_kernel_integrity(path):
+    """Check if the SPICE Kernel has the adequate architecture.
+
+    All SPICE kernels must have a NAIF file ID word as the first "word" on
+    the first line of the kernel. This "word" describes the architecture
+    of the kernel.
+
+    A binary kernel could have the following architectures:
+
+      * ``DAF``: The file is based on the DAF architecture.
+      * ``DAS``: The file is based on the DAS architecture.
+      * ``XFR``: The file is in a SPICE transfer file format.
+
+    For an archive only ``DAF`` is acceptable.
+
+    Text kernels must have ``KPL`` (Kernel Pool File) architecture.
+
+    NPB checks if binary kernels have a ``DAF`` architecture and text kernels
+    a ``KPL`` architecture.
+
+    :param path: SPICE Kernel or ORBNUM path
+    :type path: str
+    :return: Error message if error present
+    :rtype: str
+    """
+    error = ''
+    name = path.split(os.sep)[-1].strip()
+    extension = path.split(".")[-1].strip()
+    type_file = extension_to_type(name).upper()
+
+    #
+    # Determine if it is a binary or a text kernel.
+    #
+    if extension[0].lower() == "b":
+        file_format = "Binary"
+    else:
+        file_format = "Character"
+
+    #
+    # All files that are to have labels generated must have a NAIF
+    # file ID word as the first "word" on the first line of the
+    # file. Check if it is the case.
+    #
+    (arch, type) = spiceypy.getfat(path)
+
+    if file_format == "Binary":
+
+        #
+        # A binary file could have the following architectures:
+        #
+        #   DAF - The file is based on the DAF architecture.
+        #   DAS - The file is based on the DAS architecture.
+        #   XFR - The file is in a SPICE transfer file format.
+        #
+        # But for an archive only DAF is acceptable.
+        #
+        if (arch != "DAF") and (arch != "DAS"):
+            error = f"Kernel {name} architecture {arch} is invalid."
+        else:
+            pass
+
+    elif file_format == "Character":
+
+        #
+        # Text kernels must have KPL architecture:
+        #
+        #    KPL -- Kernel Pool File (i.e., a text kernel)
+        #
+        if arch != "KPL":
+            error = f"Kernel {name} architecture {arch} is invalid."
+
+    if type != type_file:
+        error = f"Kernel {name} type {type_file} is not the one expected: {type}."
+
+    return error
+
+
+def check_binary_endianness(path, endianness):
+    """Check if the SPICE Kernel has the adequate architecture.
+
+    PDS4 Bundles require LTL-IEEE binary kernels and PDS3 data sets require
[…]
