### 2022-04-21 23182c2
Updated kernel integrity tests to point to the updated functions.
