### 2022-04-25 7b01c3c
Added functions for kernel tests.
