### 2022-04-25 d2f5bef
Added functions for kernel tests.
