+++ src/pds/naif_pds4_bundler/utils/__init__.py
@@ -5,7 +5,11 @@ The utils module implements general utility capabilities used elsewhere in NPB.
 from .decorators import spice_exception_handler
 from .files import add_carriage_return
 from .files import add_crs_to_file
+from .files import check_badchar
+from .files import check_binary_endianness
 from .files import check_consecutive
+from .files import check_eol
+from .files import check_kernel_integrity
 from .files import check_list_duplicates
 from .files import checksum_from_label
 from .files import checksum_from_registry
@@ -22,15 +26,13 @@ from .files import kernel_name
 from .files import match_patterns
 from .files import md5
 from .files import mk_to_list
+from .files import product_mapping
 from .files import replace_string_in_file
 from .files import safe_make_directory
 from .files import string_in_file
 from .files import type_to_extension
 from .files import type_to_pds3_type
 from .files import utf8len
-from .files import product_mapping
-from .files import check_kernel_integrity
-from .files import check_binary_endianness
 from .time import ck_coverage
 from .time import creation_time
 from .time import current_date
@@ -81,5 +83,7 @@ __all__ = [
     format_multiple_values,
     product_mapping,
     check_kernel_integrity,
-    check_binary_endianness
+    check_binary_endianness,
+    check_badchar,
+    check_eol
 ]
