### 2022-04-25 1793d6e
Updated endianness tests assertion following implementation of kernel list checks.
