### 2022-04-25 30a2570
Updated tests to convert LF files to CLRF when required --due to GitHub feature to convert CRLF files to LF for Unix systems.
