+++ tests/naif_pds4_bundler/test_unittests.py
@@ -21,6 +21,7 @@ import unittests.test_orbnum as orbnum
 import unittests.test_plan as plan
 import unittests.test_readme as readme
 import unittests.test_time as time
+from pds.naif_pds4_bundler.utils import add_crs_to_file
 
 
 class TestUnitTests(TestCase):
@@ -51,6 +52,20 @@ class TestUnitTests(TestCase):
         tests_dir = cls.tmp_dir.name + "/naif_pds4_bundler/unittests/"
         print(f"       Tests data on: {cls.tmp_dir.name}")
         os.chdir(tests_dir)
+
+        #
+        # The ORBNUM files that are supposed to have CRLF are updated here.
+        # The reason is to avoid issues with Git LF to CRLF updates: The
+        # repository only contains files with LF line endings.
+        #
+        files = [
+            f"{tests_dir}../data/misc/orbnum/maven_orb_rec_210101_210401.orb",
+            f"{tests_dir}../data/misc/orbnum/maven_orb_rec_210101_210401_v1.orb",
+            f"{tests_dir}../data/misc/orbnum/maven_orb_rec_210101_210401_v2.orb",
+            f"{tests_dir}../data/misc/orbnum/maven_orb_rec_210101_210401_v3.orb"
+        ]
+        for file in files:
+            add_crs_to_file(file, "\r\n")
 
     @classmethod
     def tearDownClass(cls):
