### 2022-04-26 4841bc6
Enhanced kernel list check documenation and implemented a complete test case.
