+++ src/pds/naif_pds4_bundler/classes/list.py
@@ -1012,17 +1012,29 @@ class KernelList(List):
                 logging.info("")
 
     def check_products(self):
-        """Check the SPICE kernel and ORBNUM products.
+        """Check the SPICE kernel and ORBNUM products from the kernel list.
+
+        The products present in the Kernel List --whether if generated by NPB
+        or provided by the user-- will be checked and if issues are found they
+        will be reported as ``WARNING`` or ``ERROR`` messages.
 
         The checks performed to the products are the following:
 
-         * Identify if a product is present in multiple directories
-         * Check End of Line character
-         * Check files for non-printing and non-ASCII characters.
-         * Kernel architecture
-         * Endianness of binary kernels
+         * identify if a SPICE kernel is present in multiple kernel directories
+         * check End of Line character for text SPICE kernels and ORBNUM files.
+         * check text SPICE kernels and ORBNUM files for non-ASCII characters.
+         * validate kernel architecture
+         * check endianness of binary kernels
         """
-        logging.info("-- Checking kernel list products:")
+        line = f"Step {self.setup.step} - Check kernel list products"
+        logging.info("")
+        logging.info(line)
+        logging.info("-" * len(line))
+        logging.info("")
+        self.setup.step += 1
+        if not self.setup.args.silent and not self.setup.args.verbose:
+            print("-- " + line.split(" - ")[-1] + ".")
+
         #
         # A products errors list and a warnings list will be created these
         # will be stored in a product dictionary. The error list will stop
@@ -1131,12 +1143,13 @@ class KernelList(List):
 
         #
         # With all checks performed now they need to be reported.
+        # Start for logging and/or verbose mode reporting.
         #
         error_flag = False
         warning_flag = False
         for product in product_list:
             if product_errors[product] or product_warnings[product]:
-                logging.warning(f"   {product}")
+                logging.warning(f"-- {product}")
                 warning_flag = True
                 if product_warnings[product]:
                     for line in product_warnings[product]:
@@ -1146,8 +1159,28 @@ class KernelList(List):
                     for line in product_errors[product]:
                         logging.error(f"     {line}")
 
+        #
+        # Report in standard output if necessary.
+        #
+        if not self.setup.args.silent and not self.setup.args.verbose:
+            for product in product_list:
+                if product_errors[product] or product_warnings[product]:
+                    print(f"   * {product}")
+                    warning_flag = True
+                    if product_warnings[product]:
+                        for line in product_warnings[product]:
+                            print(f"     {line}")
+                    if product_errors[product]:
+                        error_flag = True
+                        for line in product_errors[product]:
+                            print(f"     {line}")
+
         if error_flag:
+            if not self.setup.args.silent and not self.setup.args.verbose:
+                print("-- Products listed above require work.")
             logging.error('')
-            error_message("Products listed above require work", self.setup)
+            error_message("Products listed above require work.", self.setup)
         elif not warning_flag:
-            logging.info("     All products checks have succeeded.")
+            logging.info("-- All products checks have succeeded.")
+            if not self.setup.args.silent and not self.setup.args.verbose:
+                print("-- All products checks have succeeded.")
