+++ src/pds/naif_pds4_bundler/classes/log.py
@@ -145,7 +145,8 @@ class Log(object):
             os.remove(template)
 
         #
-        # Generate the file list and the checksum registry.
+        # Generate the file list, the checksum registry, and the PDS validate
+        # configuration file.
         #
         self.setup.write_file_list()
         self.setup.write_checksum_registry()
