+++ docs/source/43_using_npb.rst
@@ -348,7 +348,7 @@ files. NPB's Diff files are HTML files or plain text snippets that record
 differences in between two files, highlighting differences similar to the output
 of a GUI diff tool such as ``tkdiff``. These files are useful to check and validate
 the archive increment. More information on how these files are generated is
-provided in section :ref:`43_using_npb:Validation Diff files`.
+provided in section :ref:`43_using_npb:Product Diff files`.
 
 The options for this argument determine the destination of these diff files.
 
@@ -493,7 +493,8 @@ via argument parameters. These files are:
    * Execution Log: ``<sc>_release_??.log``
    * Execution File List: ``<sc>_release_??.file_list``
    * Checksum registry:``<sc>_release_??.checksum``
-   * Validation Diff files: ``diff_<new_file>_<old_file>.html``
+   * Product Diff files: ``diff_<new_file>_<old_file>.html``
+   * PDS Validate Tool Configuration files: ``<sc>_release_??.setup.validate``
 
 where
 
@@ -703,16 +704,71 @@ The Checksum Regsitry file follows this naming schema ``<sc>_release_??.checksum
 as described in :ref:`43_using_npb:Execution by-products`.
 
 
-Validation Diff Files
-^^^^^^^^^^^^^^^^^^^^^
+Product Diff Files
+^^^^^^^^^^^^^^^^^^
 
-Validation Diff files are HTML files that provide a side to side comparison of
+Product Diff files are HTML files that provide a side to side comparison of
 a product generated by NPB and another "similar" product in order to facilitate
 checking and validating the new product, highlighting differences similar to
 the output of a GUI diff tool such as ``tkdiff``.
 
 This by-product is generated if indicated by the ``-d DIFF, --diff DIFF``
 argument.
+
+
+PDS Validate Tool Configuration File
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+NPB will write a PDS Validate Tool (``validate``) configuration file for convenience
+of the user. The options provided by the configuration file are the ones
+recommended by NAIF. Namely:
+
+   * ``verbosity`` is set to ``1`` for ``validate`` to provide all ``INFO`` level messages.
+     Some tabular data --checksum and ORBNUM products-- content checks are only
+     provided as ``INFO`` messages
+
+   * ``skip-context-validation`` is set to ``true`` and indicates validate to
+     skip the validation of the context products. This is not good practice but
+     currently the context products are not consolidated enough to take them under
+     consideration. Feel free to remove this argument or set it to ``false``
+
+   * ``rule`` set to ``true`` ensures that the bundle will be validated
+     against the schema
+
+   * ``schema``, the PDS IM Schema has to be explicitly provided to
+     ``validate`` for it to check it as it would do with any other
+     dictionary. This was not the case for older versions of validate
+
+   * ``schematron``, as per ``schema``, the schematron has to be explicitly provided
+
+   * ``strict-field-checks`` is set to ``true`` for validate to provide check the
+     tabular data --checksum and ORBNUM products--
+
+   * ``target`` simply specifies the location of the bundle
+
+   * ``report`` simply specifies the location of the resulting validation report
+
+The following ``validate`` example command for ExoMars2016::
+
+   $ validate -v 1 -t em16/em16_spice –skip-context-validation \
+   -R pds4.bundle -x working/PDS4_PDS_1B00.xsd -S working/PDS4_PDS_1B00.sch \
+   –strict-field-checks -r working/em16_release_03.validate
+
+Would be equivalent to the following resulting ``validate`` configuration file::
+
+   # Run the PDS validate tool where the NPB working directory resides:
[…]
