### 2022-04-26 e657f49
PDS Validate configuration file not to be generated for PDS3 archives.
