+++ src/pds/naif_pds4_bundler/classes/log.py
@@ -150,7 +150,8 @@ class Log(object):
         #
         self.setup.write_file_list()
         self.setup.write_checksum_registry()
-        self.setup.write_validate_configuration()
+        if self.setup.pds_version == '4':
+            self.setup.write_validate_configuration()
 
         #
         # Clear the kernel pool
