### 2022-04-26 003464b
Updated tests to prevent BIG endian file errors to be propagated.
