+++ _sources/11_introduction.rst.txt
@@ -11,6 +11,19 @@ data at some other archive facility.
 Motivation
 ==========
 
+NAIF's approach to creating SPICE kernel archives can be summarized by this
+statement:
+
+**All SPICE data for a given mission are archived as UNIX text and binary
+files in a single, accumulating archive on a single virtual volume having
+the same directory structure, the same set of meta information files, data
+file labels with the same structure, and archive documents with the same
+structure as all SPICE archives produced by NAIF.**
+
+Each time that an accumulating archive is released we either refer this to
+a release of the archive or to an archive increment. In this document you
+will find both terms used interchangeably.
+
 This document should be seen as something more than a user's software guide
 since it describes the whole process to prepare SPICE archives and it also
 describes the NAIF approach to using PDS4 standards in great detail (These
@@ -32,22 +45,6 @@ the archives. These can help a great deal, but there is much that only the
 archive preparer can do.
 
 
-NAIF's Approach to SPICE Kernel Archive Preparation
-===================================================
-
-NAIF's approach to creating SPICE kernel archives can be summarized by this
-statement:
-
-**All SPICE data for a given mission are archived as UNIX text and binary
-files in a single, accumulating archive on a single virtual volume having
-the same directory structure, the same set of meta information files, data
-file labels with the same structure, and archive documents with the same
-structure as all SPICE archives produced by NAIF.**
-
-Each time that an accumulating archive is released we either refer this to
-a release of the archive or to an archive increment. In this document you
-will find both terms used interchangeably.
-
 
 How to Read This Document?
 ==========================
@@ -66,87 +63,3 @@ dedicated to the description of functions and
 modules is aimed at potential contributors to the further development of this
 NPB software package, if you are not planning to do so, don't bother to take a
 look at that chapter.
-
-
-A Very Brief Introduction to SPICE
-==================================
-
-The SPICE (Spacecraft Planets Instrument C- Matrix Events) information
-system uses ancillary data to provide solar system geometry information
-to scientists and engineers for planetary missions in order to plan and
-analyze scientific observations from space-born instruments.
-
-A SPICE archive includes the complete set of SPICE data files
-(kernel files) for a given mission, which can be accessed using SPICE
-software. The SPICE data contain geometric and other ancillary information
-needed to recover the full value of science instrument data. In particular
-SPICE kernels provide spacecraft and planetary ephemerides,
-spacecraft and instrument orientation, instrument mounting
-alignments, data specifiying target body size, shape and orientation,
-and data needed for relevant time conversions. Data in
-SPICE kernel files must be accessed using the software called
-the SPICE Toolkit produced and distributed by the Navigation and
-Ancillary Information Facility (NAIF) Node of the Planetary Data
-System.
-
-The SPICE toolkit is available in FORTRAN, C at the
-`NAIF web site <https://naif.jpl.nasa.gov>`_.
-Interfaces to higher-level data analysis software in the Interactive Data
-Language (IDL), Matlab and JAVA are also provided. There are several other
-wrappers amongst which we highlight the Python wrapper SpiceyPy.
-
-
-SPICE Kernels Organization
---------------------------
-
-SPICE kernels in a SPICE archive are grouped into a set of subdirectories.
-Each subdirectory contains one specific kernel type; those types are briefly
-described here.
-
- * **SPK** (Spacecraft Planet Kernel) files contain ephemerides (position
-   and velocity) for spacecraft, planets, satellites, comets and
-   asteroids as well as for moving or fixed spacecraft and instrument
-   structures.
-
- * **PCK** (Planetary Constants Kernel) files contain certain physical,
-   dynamical and cartographic constants for target bodies, such as size
[…]
