### 2022-04-28 a521ab0
Updated text concerning XML Schema and Information Model parameters that are not mandatory anymore.
