### 2022-04-28 3f2e38c
Added test to generate xml_model and schema_location and adapted INSIGHT configuration file.
