+++ tests/naif_pds4_bundler/unittests/test_readme.py
@@ -102,10 +102,9 @@ def test_im_version_ascii(self):
         with open(self.updated_config, "w") as n:
             for line in c:
                 if "<information_model>1.5.0.0</information_model>" in line:
-                    n.write("<information_model>1.16.0.0" "</information_model>\n")
+                    n.write("<information_model>1.16.0.0</information_model>\n")
                 elif (
-                    "<xml_model>http://pds.nasa.gov/pds4/pds/v1/"
-                    "PDS4_PDS_1500.sch</xml_model>\n" in line
+                    "<xml_model>http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.sch</xml_model>" in line
                 ):
                     n.write(
                         "<xml_model>http://pds.nasa.gov/pds4/pds/v1/"
@@ -113,12 +112,12 @@ def test_im_version_ascii(self):
                     )
                 elif (
                     "<schema_location>http://pds.nasa.gov/pds4/pds/v1 "
-                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd" in line
+                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd<schema_location>" in line
                 ):
                     n.write(
                         "<schema_location>http://pds.nasa.gov/pds4/pds/"
                         "v1 http://pds.nasa.gov/pds4/pds/v1/"
-                        "PDS4_PDS_1G00.xsd\n"
+                        "PDS4_PDS_1G00.xsd<schema_location>\n"
                     )
                 else:
                     n.write(line)
@@ -142,7 +141,7 @@ def test_im_version_ascii_incorrect(self):
             for line in c:
                 if "<information_model>1.5.0.0</information_model>" in line:
                     n.write(
-                        "<information_model>10.16.11.20" "</information_model>\n"
+                        "<information_model>10.16.11.20</information_model>\n"
                     )
                 elif (
                     "<xml_model>http://pds.nasa.gov/pds4/pds/v1/"
@@ -154,12 +153,12 @@ def test_im_version_ascii_incorrect(self):
                     )
                 elif (
                     "<schema_location>http://pds.nasa.gov/pds4/pds/v1 "
-                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd" in line
+                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd<schema_location>" in line
                 ):
                     n.write(
                         "<schema_location>http://pds.nasa.gov/pds4/pds/"
                         "v1 http://pds.nasa.gov/pds4/pds/v1/"
-                        "PDS4_PDS_1G00.xsd\n"
+                        "PDS4_PDS_1G00.xsd<schema_location>\n"
                     )
                 else:
                     n.write(line)
