+++ docs/npb_worklist.txt
@@ -5,18 +5,6 @@ Work List:
 
   Next:
 
-  -- Generate validate tool configuration file. Setup schema download, and
-     remove the need to have these items in configuration:
-
-        <information_model>1.5.0.0</information_model>
-        <xml_model>http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.sch</xml_model>
-        <schema_location>http://pds.nasa.gov/pds4/pds/v1 http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd
-        </schema_location>
-
-     Build them up only with the IM version, they would still be optional.
-
-  -- Remove MK input from configuration file.
-
   -- Implement OREX functional tests -focusing on MKs and kernel mapping.
 
   -- Implement BINGO dependency for LTL-IEEE machines for PDS3 data sets.
@@ -25,9 +13,6 @@ Work List:
 
   -- Include option to provide plan with kernel names to be mapped or to
      provide kernels from directory with original names.
-
-  -- Agreed to try to enhance the MK generation with whitespaces
-     for selected kernels of the same type.
 
   -- New product for PDS3 archives: MKs in data/mk. To be a copy of the
      extras/mk products with _pds4_ added to the filename with a PATH value
@@ -47,7 +32,24 @@ Work List:
 
 Changelog:
 
-v1.0.3
+
+
+
+v1.1.1
+
+  -- Remove the need to have these items in configuration:
+
+        <information_model>1.5.0.0</information_model>
+        <xml_model>http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.sch</xml_model>
+        <schema_location>http://pds.nasa.gov/pds4/pds/v1 http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd
+        </schema_location>
+
+     Build them up only with the IM version, they would still be optional.
+
+  -- Agreed to try to enhance the MK generation with whitespaces
+     for selected kernels of the same type.
+
+v1.1.0
 
   -- Advance kernel checks to kernel list generation including:
 
@@ -58,6 +60,8 @@ v1.0.3
 
      The documentation has to be updated accordingly. Also document the fact
      that the "list" faucet will not run the checks.
+
+  -- Generate validate tool configuration file. Setup schema download
 
 v0.19.0
 
