+++ docs/source/42_npb_configuration_file.rst
@@ -964,8 +964,8 @@ of the required and optional directories:
      - Yes
    * - kernels_directory
      - Specifies the directory(ies) that will be used by NPB to obtain the kernels
-       to be archived from. These directories must follow the usual operational
-       SPICE kernel sub-directory structure by kernel type.
+       to be archived from. It is recommended that these directories follow the
+       usual operational SPICE kernel sub-directory structure by kernel type.
      - Yes
    * - staging_directory
      - This directory will be used by NPB to store the files generated by its
@@ -983,7 +983,8 @@ of the required and optional directories:
      - Indicates the directory where the user input templates are present.
      - No
 
-More information on the setup of the NPB directories is provided in
+The working, staging, and bundle directories cannot be the same. More
+information on the setup of the NPB directories is provided in
 :ref:`32_step_2_npb_setup:Workspace Setup`.
 
 
