### 2022-04-29 bf546ad
Fixed tests to incorporate kernel directory and IM models checks.
