+++ tests/naif_pds4_bundler/unittests/test_readme.py
@@ -30,7 +30,7 @@ def test_im_format(self):
         with open(self.updated_config, "w") as n:
             for line in c:
                 if "<information_model>1.5.0.0</information_model>" in line:
-                    n.write("<information_model>1.A.0.0" "</information_model>\n")
+                    n.write("<information_model>1.A.0.0</information_model>\n")
                 else:
                     n.write(line)
 
@@ -75,20 +75,16 @@ def test_im_schema_incoherent(self):
     with open(self.config, "r") as c:
         with open(self.updated_config, "w") as n:
             for line in c:
-                if (
-                    "<schema_location>http://pds.nasa.gov/pds4/pds/v1 "
-                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd" in line
-                ):
+                if ("PDS4_PDS_1500.xsd" in line):
                     n.write(
-                        "<schema_location>http://pds.nasa.gov/pds4/pds/"
-                        "v1 http://pds.nasa.gov/pds4/pds/v1/"
-                        "PDS4_PDS_1A00.xsd\n"
+                        "<schema_location>http://pds.nasa.gov/pds4/pds/ v1 "
+                        "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1A00.xsd</schema_location>\n"
                     )
                 else:
                     n.write(line)
 
     with self.assertRaises(RuntimeError):
-        main(self.updated_config, faucet=self.faucet, silent=True)
+        main(self.updated_config, faucet=self.faucet, silent=True, log=True)
 
 
 def test_im_version_ascii(self):
@@ -112,12 +108,12 @@ def test_im_version_ascii(self):
                     )
                 elif (
                     "<schema_location>http://pds.nasa.gov/pds4/pds/v1 "
-                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd<schema_location>" in line
+                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd</schema_location>" in line
                 ):
                     n.write(
                         "<schema_location>http://pds.nasa.gov/pds4/pds/"
                         "v1 http://pds.nasa.gov/pds4/pds/v1/"
-                        "PDS4_PDS_1G00.xsd<schema_location>\n"
+                        "PDS4_PDS_1G00.xsd</schema_location>\n"
                     )
                 else:
                     n.write(line)
@@ -153,12 +149,12 @@ def test_im_version_ascii_incorrect(self):
                     )
                 elif (
                     "<schema_location>http://pds.nasa.gov/pds4/pds/v1 "
-                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd<schema_location>" in line
+                    "http://pds.nasa.gov/pds4/pds/v1/PDS4_PDS_1500.xsd</schema_location>" in line
                 ):
                     n.write(
                         "<schema_location>http://pds.nasa.gov/pds4/pds/"
                         "v1 http://pds.nasa.gov/pds4/pds/v1/"
-                        "PDS4_PDS_1G00.xsd<schema_location>\n"
+                        "PDS4_PDS_1G00.xsd</schema_location>\n"
                     )
                 else:
                     n.write(line)
