### 2022-04-29 85401ee
Fixed tests to incorporate kernel directory and IM models checks.
