+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py39, docs, lint
+envlist = py38, docs, lint
 
 [testenv]
 deps = .[dev]
@@ -21,6 +21,6 @@ commands=
 skip_install = true
 
 [testenv:dev]
-basepython = python3.9
+basepython = python3.8
 usedevelop = True
 deps = .[dev]
