### 2022-05-02 a0ae8ff
Moved generation of run by-products in order for them to be present in the run log.
