### 2022-05-02 d465a82
Moved generation of run by-products in order for them to be present in the run log.
