+++ setup.cfg
@@ -25,6 +25,7 @@ install_requires =
     nose
     coverage
     xmlschema
+    requests
 # Change this to False if you use things like __file__ or __path__—which you
 # shouldn't use anyway, because that's what ``pkg_resources`` is for 🙂
 zip_safe = False
