### 2022-05-02 1ea770e
Added enhanced logging for incorrect equal working, staging or bundle directories.
