### 2022-05-04 9eddbfe
Fixed issue with copying files from multiple kernel directories. Recovered addition of CRLF in ORBNUM files for PDS4.
