+++ src/pds/naif_pds4_bundler/classes/product.py
@@ -1637,7 +1637,7 @@ class OrbnumFileProduct(Product):
         #
         if self.setup.pds_version == '4':
             if check_eol(product_path + os.sep + self.name, self.setup.eol):
-                logging.info("--Adding CRLF to ORBNUM file.")
+                logging.info("-- Adding CRLF to ORBNUM file.")
                 add_crs_to_file(product_path + os.sep + self.name, self.setup.eol, self.setup)
 
         #
