### 2022-05-11 #4327 @mbobra
👋  @mcostasitja Let's see what happens with the reviewers' time schedules. Remember they review for JOSS on volunteer time, so they may not be able to complete it before May 19. If that's the case and you need to withdraw, don't worry -- JOSS does not penalize in any way for withdrawing and resubmitting. We can basically re-assign this submission right away, I will be happy to act as the editor again, and you won't lose any time.
