### 2022-05-11 #4327 @mcostasitja
@mbobra, @SimonMolinsky thanks a lot for getting back to me promptly.

I've been a reviewer multiple times for different Journals, all on volunteer time -the perk is to access to the Journal but it's common to have institutional access already- so I am very empathic towards reviewers; also there have always been deadlines for me as a reviewer.

I submitted the paper one month ago and my affiliation change was not certain at the time. One of the main reasons for me to submit this paper was to be able to have it with my current affiliation, if that does not happen, it makes little sense to maintain the submission because it has been developed under that affiliation.

I'll check with my management whether if somehow the reviewed contribution can be approved a posteriori but is highly unlikely.
