### 2022-05-11 #4327 @athulpg007
I tried running the tests and got this, did I miss something?

python -m unittest
EEE
ERROR: naif_pds4_bundler.test_functional (unittest.loader._FailedTest)
ImportError: Failed to import test module: naif_pds4_bundler.test_functional
Traceback (most recent call last):
  File "/usr/lib/python3.8/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/usr/lib/python3.8/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/athul/Documents/Codes/naif-pds4-bundler/tests/naif_pds4_bundler/test_functional.py", line 8, in <module>
    import functional.test_dart as dart
ModuleNotFoundError: No module named 'functional'

ERROR: naif_pds4_bundler.test_regression (unittest.loader._FailedTest)
ImportError: Failed to import test module: naif_pds4_bundler.test_regression
Traceback (most recent call last):
  File "/usr/lib/python3.8/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/usr/lib/python3.8/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/athul/Documents/Codes/naif-pds4-bundler/tests/naif_pds4_bundler/test_regression.py", line 8, in <module>
    import regression.test_pds3 as pds3
ModuleNotFoundError: No module named 'regression'

ERROR: naif_pds4_bundler.test_unittests (unittest.loader._FailedTest)
ImportError: Failed to import test module: naif_pds4_bundler.test_unittests
Traceback (most recent call last):
  File "/usr/lib/python3.8/unittest/loader.py", line 436, in _find_test_path
    module = self._get_module_from_name(name)
  File "/usr/lib/python3.8/unittest/loader.py", line 377, in _get_module_from_name
    __import__(name)
  File "/home/athul/Documents/Codes/naif-pds4-bundler/tests/naif_pds4_bundler/test_unittests.py", line 9, in <module>
    import unittests.test_bundle_history as bundle_history
ModuleNotFoundError: No module named 'unittests'

Ran 3 tests in 0.000s
FAILED (errors=3)
