### 2022-05-11 #4327 @mcostasitja
> I tried running the tests and got this, did I miss something?
>
> ```
> python -m unittest

Have you tried under the tests/naif_pds4_bundler directory? This is indicated here: https://nasa-pds.github.io/naif-pds4-bundler/41_npb_installation.html#running-tests
