### 2022-05-12 #4327 @athulpg007
@mcostasitja I have completed my first review of NAIF PDS4 Bundler. This is a well-developed and well-documented package which has demonstrated its usefulness to generate PDS4 archives within NASA, ESA, and JAXA. I have gone through  the checklist and checked off all items except one, which I have minor comment about.

In the documentation, would it be possible to add at least one example (could be simple) of how a user would use the package to generate an archive? I think it this would greatly enhance the usability of this package for someone who is not familiar with it in their day-to-day work.
