### 2022-05-12 #4327 @mcostasitja
@athulpg007 thanks for taking the time to do this first round. I fully agree with your suggestion and I will include an example in the documentation. More specifically in after the "Using NPB" section: https://nasa-pds.github.io/naif-pds4-bundler/43_using_npb.html#

I will call it "Working Example" -suggestions are welcome, I am not an English native speaker. The example will use existing files present in the data tests directories. The example will use a fix version of NPB. This will be included in a release that I will push by the end of the week. Hopefully tomorrow.
