+++ docs/source/44_working_example.rst
@@ -1,1 +1,3 @@
 This will be a working example.
+
+Example.
