+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [v1.1.1-dev](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.1.1-dev) (2022-05-04)
+## [v1.1.1-dev](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.1.1-dev) (2022-05-12)
 
 [Full Changelog](https://github.com/NASA-PDS/naif-pds4-bundler/compare/v1.1.0...v1.1.1-dev)
 
