### 2022-05-12 #4327 @mcostasitja
@athulpg007 @SimonMolinsky @mbobra I just checked with my management and it looks like it will be possible to clear the final version of the article even if I am not longer affiliated to NASA/JPL so we are good to go for a review ending later.
