+++ docs/source/40_naif_pds4_bundler_user_guide.rst
@@ -11,3 +11,4 @@ understand the NAIF PDS4 Bundler Software Package.
    41_npb_installation
    42_npb_configuration_file
    43_using_npb
+   44_working_example
