+++ docs/source/44_working_example.rst
@@ -1,3 +1,184 @@
-This will be a working example.
+Working Example
+===============
 
-Example.
+This section provides a NPB working example for the user to be able to exercise
+NPB's execution before using it with actual data. This section assumes that
+NPB is properly installed as indicated in :ref:`41_npb_installation:NPB Installation`
+and that you have the package available in your compute. We will assume that
+its location is ``~/naif-pds4-bundler/`` and that you can execute NPB with
+``$naif-pds4-bundler``.
+
+This working example has been generated with NPB v1.2.0.
+
+Setup the workspace
+-------------------
+
+We will generate an increment of the INSIGHT archive. The first step is to
+define our workspace, we will do so by creating the adequate directory
+structure as follows::
+
+    cd ~/
+    mkdir insight
+    mkdir insight/working
+    mkdir insight/staging
+    mkdir insight/bundle
+    mkdir insight/kernels
+    mkdir insight/kernels/sclk
+    mkdir insight/kernels/ck
+
+The next step is to copy the "current" version of the INSIGHT SPICE Kernel
+(version 7). This is provided as test data with NPB::
+
+    cp -pr naif-pds4-bundler/tests/naif_pds4_bundler/data/insight/insight_spice insight/bundle/
+
+Now we copy the kernels that we will include in the archive increment (that will
+be version 8)::
+
+   cp -p naif-pds4-bundler/tests/naif_pds4_bundler/data/kernels/sclk/NSY_SCLKSCET.00019.tsc insight/kernels/sclk/nsy_sclkscet_00019.tsc
+   cp -p naif-pds4-bundler/tests/naif_pds4_bundler/data/kernels/ck/insight_ida_enc_200829_201220_v1.bc insight/kernels/ck/
+   cp -p naif-pds4-bundler/tests/naif_pds4_bundler/data/kernels/ck/insight_ida_pot_200829_201220_v1.bc insight/kernels/ck/
+
+The next step is to write the archive release plan that indicates the SPICE
+kernels that we want to archive. We will also include the meta-kernel in the
+release plan -- it will be automatically generated by NPB and therefore
+we do not need to provide it. You can generate the release plan as follows::
+
+   {ls -1 insight/kernels/sclk; ls -1 insight/kernels/ck} > insight/working/insight_release_08.plan
+
+The next step is to setup the NPB configuration file appropriately. The starting
+point will be the configuration file provided for INSIGHT in the NPB test data::
+
+   cp naif-pds4-bundler/tests/naif_pds4_bundler/config/insight.xml insight/working/insight_release_v08.xml
+
+
+Prepare the Configuration File
+------------------------------
+
+We will now need to modify the configuration file to adapt it to our workspace.
+In order to do so open the file with your favorite editor and do the replacements
+indicated hereunder::
+
+   vi insight/working/insight_release_v08.xml
+
+Because we are not providing an updated SPICEDS file, remove the following line::
+
+   <spiceds>../data/spiceds_insight.html</spiceds>
+
+Replace::
+
+   <sclk>NSY_SCLKSCET.[0-9][0-9][0-9][0-9][0-9].tsc</sclk>
+
+with::
+
+   <sclk>nsy_sclkscet_[0-9][0-9][0-9][0-9][0-9].tsc</sclk>
+
+Given that we do not need to map the SCLK kernel name -- we already copied the
+file with the update name.
+
+Replace::
+
+   <bundle_directory>insight</bundle_directory>
+
+with::
+
+   <bundle_directory>bundle</bundle_directory>
+
+
+Finally, replace::
+
+   <interrupt_to_update>True</interrupt_to_update>
+
+with::
+
+   <interrupt_to_update>False</interrupt_to_update>
+
[…]
