### 2022-05-14 1c8b5b6
Fixed bug with interactive MK parameter set to False.
