### 2022-05-14 e578aff
Inspect Bundle directory for kernels to load in addition to kernels directory.
