+++ docs/source/44_working_example.rst
@@ -1,14 +1,13 @@
 Working Example
 ===============
 
-This section provides a NPB working example for the user to be able to exercise
-NPB's execution before using it with actual data. This section assumes that
-NPB is properly installed as indicated in :ref:`41_npb_installation:NPB Installation`
-and that you have the package available in your compute. We will assume that
-its location is ``~/naif-pds4-bundler/`` and that you can execute NPB with
-``$naif-pds4-bundler``.
+This section provides a working example for the user to be able to exercise
+NPB's execution before using it to generate an actual archive. This section
+assumes that NPB is properly installed as indicated in :ref:`41_npb_installation:NPB Installation`
+and that you have the package available in the user's computer. We will assume that
+its location is ``~/naif-pds4-bundler/`` and that NPB can executed with the
+command ``naif-pds4-bundler``.
 
-This working example has been generated with NPB v1.2.0.
 
 Setup the workspace
 -------------------
@@ -41,7 +40,7 @@ be version 8)::
 The next step is to write the archive release plan that indicates the SPICE
 kernels that we want to archive. We will also include the meta-kernel in the
 release plan -- it will be automatically generated by NPB and therefore
-we do not need to provide it. You can generate the release plan as follows::
+we do not need to provide it. The release plan can be generated as follows::
 
    {ls -1 insight/kernels/sclk; ls -1 insight/kernels/ck} > insight/working/insight_release_08.plan
 
@@ -55,7 +54,7 @@ Prepare the Configuration File
 ------------------------------
 
 We will now need to modify the configuration file to adapt it to our workspace.
-In order to do so open the file with your favorite editor and do the replacements
+In order to do so open the file with a text editor and do the replacements
 indicated hereunder::
 
    vi insight/working/insight_release_v08.xml
@@ -92,15 +91,15 @@ with::
 
    <interrupt_to_update>False</interrupt_to_update>
 
-This will prevent NPB from pausing the process for you to modify the
-automatically generated meta-kernel.
+This will prevent NPB from pausing the process to update the automatically
+generated meta-kernel.
 
 
 Run NPB
 -------
 
 We are ready to run NPB and to generate the archive increment. To do so change
-the directory to ``insight`` and then run NPB as follows::
+the directory to ``insight`` and then run NPB::
 
    cd insight
    naif-pds4-bundler working/insight_release_v08.xml -p working/insight_release_08.plan -l
@@ -172,8 +171,8 @@ happen due to the usage of different versions of NPB)::
 Checking the Results
 --------------------
 
-You are all set. You can check the result by inspecting the run by-products if
-you wish, these are under the ``working`` directory::
+NPB has been executed successfully. The result can be checked by inspecting the
+run by-products that are under the ``working`` directory::
 
    PDS4_PDS_1500.sch
    PDS4_PDS_1500.xsd
