### 2022-05-16 5edf303
Corrected flake8 findings.
