+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [v1.2.0-dev](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.2.0-dev) (2022-05-14)
+## [v1.2.0-dev](https://github.com/NASA-PDS/naif-pds4-bundler/tree/v1.2.0-dev) (2022-05-16)
 
 [Full Changelog](https://github.com/NASA-PDS/naif-pds4-bundler/compare/v1.1.0...v1.2.0-dev)
 
