+++ _sources/42_npb_configuration_file.rst.txt
@@ -459,13 +459,13 @@ products. The table below provides a summary of the parameters:
    * - xml_model
      - URL location of the XML Schematron associated with an information model.
        The ``information_model`` and ``xml_model`` parameters must refer to the
-       same information model.
-     - Yes
+       same information model. If not provided NPB generates it from the ``information_model``.
+     - No
    * - schema_location
      - URL location of the XML Schema associated with an information model. The
        ``schema_location`` and ``xml_model`` parameters must refer to the same
-       information model.
-     - Yes
+       information model. If not provided NPB generates it from the ``information_model``.
+     - No
    * - logical_identifier
      - Logical identifier for the bundle.
      - Yes
@@ -964,8 +964,8 @@ of the required and optional directories:
      - Yes
    * - kernels_directory
      - Specifies the directory(ies) that will be used by NPB to obtain the kernels
-       to be archived from. These directories must follow the usual operational
-       SPICE kernel sub-directory structure by kernel type.
+       to be archived from. It is recommended that these directories follow the
+       usual operational SPICE kernel sub-directory structure by kernel type.
      - Yes
    * - staging_directory
      - This directory will be used by NPB to store the files generated by its
@@ -983,7 +983,8 @@ of the required and optional directories:
      - Indicates the directory where the user input templates are present.
      - No
 
-More information on the setup of the NPB directories is provided in
+The working, staging, and bundle directories cannot be the same. More
+information on the setup of the NPB directories is provided in
 :ref:`32_step_2_npb_setup:Workspace Setup`.
 
 
@@ -1513,10 +1514,12 @@ Here's an example for InSight::
 Manual Input
 ^^^^^^^^^^^^
 
-If you choose to provide a manually generated MK, you only need to provide
-the location of the MK(s) using the required number of entries of the
-``file`` element in ``mk_inputs``. E.g. for the LADEE archive the MK section
-of the configuration file can be as simple as: ::
+If you choose to provide a manually generated MK, an alternative to specify the
+MK in the release plan or in the kernel list is to provide it in the MK section
+of the configuration file: you can provide the location of the MK(s) using the
+required number of entries of the ``file`` element in ``mk_inputs``. E.g. for
+the LADEE archive the MK section of the configuration file can be as simple
+as::
 
     <meta-kernel>
         <mk_inputs>
@@ -1528,6 +1531,9 @@ of the configuration file can be as simple as: ::
             </name>
          </mk>
     </meta-kernel>
+
+If the MKs specified by this list are not found the pipeline will stop its
+execution.
 
 
 Coverage determination
