### 2022-05-16 #4327 @marcsit
@athulpg007 I have released a new version of the package that includes the Working Example in the documentation:
https://nasa-pds.github.io/naif-pds4-bundler/44_working_example.html

The assets are available at: https://github.com/NASA-PDS/naif-pds4-bundler/releases/tag/v1.2.0
