### 2022-05-22 #4327 @athulpg007
Thanks for adding the working example.

Some minor suggestions to improve the docs I have are listed below. If you agree with them, you can consider making these changes.

1. The "installation" instructions are nested within the User Guide section in the table of contents sidebar. Since the installation page is something most people would look for first, I suggest moving this out to its own section below "Overview", or Kernel Preparation Archive Guide.
2. I suggest highlighting "Your computer must be based on a 64-bit Unix operating system: a Linux or OSX." in the installation instructions as a "warning" to users who might not be one of these systems.
3. Suggest moving the "Contribution Instructions" out of the installation page, and into its own section in the table of contents.
4. Suggest renaming "Modules, Classes, and Functions Documentation" to "API Reference", since this is something developers look for and is more concise.
5. Suggest adding a section in the documentation with the links to some of the archives produced using NAIF PDS4 bundler eg: Mars 2020, MAVEN etc. to highlight the practical use of the package. I feel this would be helpful to someone who may read the documentation page, but may not have read the paper.

I have completed my review and am happy to accept this submission.
