### 2022-06-05 #4327 @SimonMolinsky
Hi @marcsit ,

In the beginning, thank you for your work! The use of `naif-pds4-bundler` is simple, and it works fast & fine. You've created thorough documentation, and it guarantees reusability in the future.

The paper is clear, and I understood the scope of the software at the first reading.

I like the test coverage of the code, it is extensive and logically sorted. The only glitches related to code are those in the comment above. Please, refer to it. Coming back to the tests, I've seen that you've provided documentation for the test classes/cases here: https://nasa-pds.github.io/naif-pds4-bundler/tests.html, but it is corrupted. I would rather skip tests documentation, but if you want to, then take a look into your Sphinx settings. I ran the example from the conda environment, and the package worked without problems.
