### 2022-06-07 d0fcf39
Updated tox to be Python3 version agnostic.
