+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py38, docs, lint
+envlist = py3, docs, lint
 
 [testenv]
 deps = .[dev]
