### 2022-06-07 #4327 @marcsit
@athulpg007 thanks a lot for your review. Your comments are highly appreciated!

All the feedback that I incorporate as per the comments below will be available in the upcoming release v1.2.1

Replying to your suggestions:

> The "installation" instructions are nested within the User Guide section in the table of contents sidebar. Since the installation page is something most people would look for first, I suggest moving this out to its own section below "Overview", or Kernel Preparation Archive Guide.

Because of the wider scope of the documentation --which might not be the best approach-- instead I have added a note and a pointer to the installation instructions in the "How to read this document" section: https://nasa-pds.github.io/naif-pds4-bundler/11_introduction.html#how-to-read-this-document

> I suggest highlighting "Your computer must be based on a 64-bit Unix operating system: a Linux or OSX." in the installation instructions as a "warning" to users who might not be one of these systems.

Highlighted.

> Suggest moving the "Contribution Instructions" out of the installation page, and into its own section in the table of contents.

Moved.

> Suggest renaming "Modules, Classes, and Functions Documentation" to "API Reference", since this is something developers look for and is more concise.

Renamed. I was never happy with that title.

> Suggest adding a section in the documentation with the links to some of the archives produced using NAIF PDS4 bundler eg: Mars 2020, MAVEN etc. to highlight the practical use of the package. I feel this would be helpful to someone who may read the documentation page, but may not have read the paper.

Added "PDS4 SPICE kernel archives examples" section.
