### 2022-06-07 #4327 @marcsit
@SimonMolinsky thanks a lot for your review, it was worth the wait! -and there is no rush anyway.

> I've checked the code, and there are multiple potential problems, but first of all, in [Contributing guidelines](https://github.com/NASA-PDS/.github/blob/main/CONTRIBUTING.md), you have a text:
>
> ```
> -   **Is it a cosmetic change, such as code formatting, white space, etc.?** We generally don't use these kinds of contributions.
>     -   Cosmetic changes create unnecessary work in reviewing them as well as noise in the commit history.
>     -   They also make backporting bugs a lot more difficult.
> ```
> I understand the reason behind it, but in the case of Python (where the syntax is essential), it could lead to problems. What kind of problems? Sometimes it is hard to draw a line between a bug/vulnerability/code patch and a syntax inconsistency with PEP8. Maybe you should update your guidelines with this information (about PEP8) to avoid a too broad interpretation of this paragraph?

I have reported the feedback to the PDS Engineering Node who are responsible of these guidelines which are "out of the scope" of NPB. I will CC you on the email I will send them.

> ## Back to the code
> When I opened the package in PyCharm Professional, then the code inspection tool showed me multiple warnings (medium level) and a few week warnings. We can group them into those classes:
>
> * `init` method should return `None` (`object` is returned),
> * A class inheritance from `object` - in Python 3.X, it is the default behavior, and I never saw this kind of inheritance in any code in Python 3.X. What was the reason behind this decision? It makes code less readable and confuses code checking and code completion.

Probably my lack of a deeper understanding of Python, I followed some IDE suggestions. My bad.

> * Definition of instance attributes OUTSIDE the `init` method, it is tough to track all class attributes in this case, and code is very sensitive to changes,
> * Too broad exception (`BaseException`) - again, it makes code hard to debug,
> * Some variables are not used,
> * and multiple cases where built-in names are shadowed: list, format, type, bytes, dir, id.
>
> Those problems are not critical from the user's perspective, but they introduce a technical debt that will slow down the development process. If you give me the green light, I can quickly refactor those code fragments and send you multiple PRs - one per problem. If you want to do it by yourself, I can open multiple Issues for you.

Please feel free to do so via Pull Request. It would be extremely useful.
