### 2022-06-08 fd8adda
Updated documentation as per JOSS review.
