### 2022-06-10 #4327 @marcsit
@SimonMolinsky NASA PDS developers have confirmed that they will add the suggestion to include the PEP8 exception to their contribution recommendations.
