### 2022-06-11 #4327 @SimonMolinsky
Hi @marcsit,

I'll start from the last comment:

> NASA PDS developers have confirmed that they will add the suggestion to include the PEP8 exception to their contribution recommendations.

It is great to hear!

> Please feel free to do so via Pull Request. It would be extremely useful.

I'm happy that my review was valuable to you. And I'll work on those issues this weekend. Hopefully, it will be a fast ride (with the test coverage you've written, it should be). I'll let you know in this thread when I'm done.
