### 2022-06-16 NASA-PDS/naif-pds4-bundler#17 @SimonMolinsky
JOSS Review: Python-related code issues
## 💡 Description
The code within the package is inconsistent with Python 3.X, and there are several medium and weak issues. Here's the list of things to correct,

- [x] init method should return None (object is returned),
- [ ] A class inheritance from the `object` - not required in Python 3.X.
- [ ] Definition of instance attributes OUTSIDE the init method, it is tough to track all class attributes in this case, and code is susceptible to changes - define attributes in the `__init__` as a `None`,
- [ ] Some variables are not used,
- [ ] Multiple cases where built-in names are shadowed: list, format, type, bytes, dir, id,
- [ ] Class `__init__` method parameters type is defined as `object`, but it can be defined strictly as a specific Class.

URL to JOSS review: https://github.com/openjournals/joss-reviews/issues/4327
