### 2022-06-16 077b88f
init returns nothing (joss rev)
