### 2022-06-16 d269eba
removed object inheritance from class constructor
