+++ src/pds/naif_pds4_bundler/classes/label.py
@@ -16,7 +16,7 @@ from ..utils import type_to_pds3_type
 from .log import error_message
 
 
-class PDSLabel(object):
+class PDSLabel():
     """Class to generate a PDS Label.
 
     :param setup: NPB execution Setup object
