### 2022-06-16 NASA-PDS/naif-pds4-bundler#17 @SimonMolinsky
`Definition of instance attributes OUTSIDE the init method, it is tough to track all class attributes in this case, and code is susceptible to changes - define attributes in the __init__ as a None`

- the task is related to the exceptions that are an important part of a process, it requires a lot of refactoring, so for now it is halted.
