### 2022-06-16 6f4a803
small typos; wrong variables not detected
