### 2022-06-16 12b26c0
updated typing in classes
