### 2022-06-16 #4327 @SimonMolinsky
Hi @marcsit , Hi @mbobra

I've updated your code @marcsit . Along the way, I've identified more things that could be improved (as always with a code :) ), but this is outside the scope of JOSS and doesn't affect performance and usability. I like how you've designed tests in the package; it helps with refactoring.

Anyway, I'm convinced that the package works well. You have my green light; @mbobra - the paper is free to go :) Thank you for your hard work!
