### 2022-06-16 NASA-PDS/naif-pds4-bundler#17 @SimonMolinsky
Hi @marcsit ,

I've made a PR into the package repo here: https://github.com/NASA-PDS/naif-pds4-bundler/pull/18

If you need more info, just let me know; the package is accepted from my side, and now it's only a matter of improvement. As you maybe see in the comment above, there is a problem with class attributes (those with `self.`) defined outside the `__init__` method. Your program catches exceptions where those attributes are undefined, and this is fine because it works, but it could be improved to be more "Pythonic." This issue is related to another problem - with broad exceptions. Those are very dangerous; we can catch something we don't want to, and we won't know about it; this is an antipattern in Python.

There are a few more issues, but too much for one thread is not a good idea :)

So just in case: will you be a maintainer in the future? Will there be anybody to accept Issues / check PRs?
