### 2022-06-17 #4327 @mbobra
Congratulations @marcsit  -- we're almost ready to publish the paper!

Could you please cut a new release of NAIF PDS4 Bundler, deposit it in Zenodo, and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
