### 2022-07-01 941628f
Updated logging level depending of number of expected kernels in MK.
