### 2022-07-05 96e50ff
DSK files were treated for validation as DAF files but they are DAS files. Logic has been added to distinguish in between them and use the adequate API.
