### 2022-07-05 33c627c
Inerrupt to update keyword is not mandatory as per configuration so the logic to consider it false has been added.
