+++ docs/requirements/v1.2.1/REQUIREMENTS.md
@@ -1,0 +1,42 @@
+
+Requirements Summary
+====================
+
+# default
+
+## Check ASCII characters in kernel list in the configuration file to avoid  the VCO IK description issue. ([#5](https://github.com/NASA-PDS/naif-pds4-bundler/issues/5)) 
+
+
+This requirement is not impacted by the current version
+## Log all the kernels used to determine MK coverage, regardless if they are in the plan or already in the bundle. ([#6](https://github.com/NASA-PDS/naif-pds4-bundler/issues/6)) 
+
+
+This requirement is not impacted by the current version
+## If specified MKs are not found, stop the pipeline. ([#7](https://github.com/NASA-PDS/naif-pds4-bundler/issues/7)) 
+
+
+This requirement is not impacted by the current version
+## Link Git repository in documentation whenever necessary ([#10](https://github.com/NASA-PDS/naif-pds4-bundler/issues/10)) 
+
+
+This requirement is not impacted by the current version
+## Do not generate validate config file for a clear run. ([#11](https://github.com/NASA-PDS/naif-pds4-bundler/issues/11)) 
+
+
+This requirement is not impacted by the current version
+## Check line length of kernels ([#12](https://github.com/NASA-PDS/naif-pds4-bundler/issues/12)) 
+
+
+This requirement is not impacted by the current version
+## Remove strict order of directories in configuration. ([#13](https://github.com/NASA-PDS/naif-pds4-bundler/issues/13)) 
+
+
+This requirement is not impacted by the current version
+## Implement OSIRIS ReX functional tests  ([#15](https://github.com/NASA-PDS/naif-pds4-bundler/issues/15)) 
+
+
+This requirement is not impacted by the current version
+## Allow comment syntax for release plan ([#16](https://github.com/NASA-PDS/naif-pds4-bundler/issues/16)) 
+
+
+This requirement is not impacted by the current version
