### 2022-07-06 NASA-PDS/naif-pds4-bundler#17 @marcsit
@SimonMolinsky I am going to be the main maintainer for a while although is not part of my job anymore. There should be a person from NAIF/JPL, the job announcement is now posted (my replacement.) And they should become the main maintainer. For the time being I will assign this issue to myself for the class attributes issues.
