+++ tests/naif_pds4_bundler/unittests/test_endianness.py
@@ -17,18 +17,16 @@ def post_setup(self):
 
     This method will be executed before each test function.
     """
-    dirs = ["kernels/fk",
-            "kernels/lsk",
-            "kernels/spk",
-            "kernels/mk"
-            ]
+    dirs = ["kernels/fk", "kernels/lsk", "kernels/spk", "kernels/mk"]
     for dir in dirs:
         os.mkdir(dir)
 
     shutil.copy2("../data/kernels/lsk/naif0012.tls", "kernels/lsk/")
     shutil.copy2("../data/kernels/mk/m2020_v01.tm", "kernels/mk/")
     shutil.copy2("../data/kernels/mk/m2020_chronos_v01.tm", "kernels/mk/")
-    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp", "kernels/spk/")
+    shutil.copy2(
+        "../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp", "kernels/spk/"
+    )
 
 
 def test_pds4_big_endianness(self):
@@ -41,8 +39,10 @@ def test_pds4_big_endianness(self):
 
     config = "../config/mars2020.xml"
 
-    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
-                 "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp")
+    shutil.copy2(
+        "../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
+        "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp",
+    )
 
     try:
         main(config, silent=True, log=True, faucet="Bundle")
@@ -63,7 +63,7 @@ def test_pds4_big_endianness_config(self):
     post_setup(self)
 
     config = "../config/mars2020.xml"
-    updated_config = 'working/mars2020.xml'
+    updated_config = "working/mars2020.xml"
     with open(config, "r") as c:
         with open(updated_config, "w") as n:
             for line in c:
@@ -73,8 +73,10 @@ def test_pds4_big_endianness_config(self):
                 else:
                     n.write(line)
 
-    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
-                 "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp")
+    shutil.copy2(
+        "../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.big.bsp",
+        "kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp",
+    )
 
     main(updated_config, silent=True, log=True, faucet="Bundle")
 
@@ -93,12 +95,15 @@ def test_pds4_ltl_endianness(self):
 
     config = "../config/mars2020.xml"
 
-    shutil.copy2("../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp",
-                 "kernels/spk/")
+    shutil.copy2(
+        "../data/kernels/spk/m2020_surf_rover_loc_0000_0089_v1.bsp", "kernels/spk/"
+    )
 
     main(config, silent=True, log=True)
 
-    line_check = "Binary SPICE kernels expected to have LTL-IEEE (little endian) binary format"
+    line_check = (
+        "Binary SPICE kernels expected to have LTL-IEEE (little endian) binary format"
+    )
     if not string_in_file("working/mars2020_release_01.log", line_check, 1):
         raise BaseException
 
@@ -108,7 +113,7 @@ def test_pds4_ltl_endianness_config(self):
     post_setup(self)
 
     config = "../config/mars2020.xml"
[…]
