### 2022-07-06 3a14e45
Removed mypy to by-pass checks.
