### 2022-07-07 NASA-PDS/naif-pds4-bundler#17 @SimonMolinsky
@marcsit I've asked because I think I can do something good here, but I'd rather wait for the publication in JOSS to avoid any potential delays. I hope that your work will be published soon! After publication, I'll "activate myself" here and hopefully, we will be able to delegate some small fixes for me (?)
