### 2022-07-07 NASA-PDS/naif-pds4-bundler#17 @marcsit
@SimonMolinsky absolutely! Thanks for that.
