### 2022-07-07 #4327 @danielskatz
Suggested changes are in https://github.com/NASA-PDS/naif-pds4-bundler/pull/21 - please merge this, or let me know what you disagree with, then we can proceed.
