### 2022-07-07 #4327 @mbobra
Thank you both so very much for your thorough reviews, @athulpg007 and @SimonMolinsky! Your volunteer time keeps JOSS running and the JOSS team appreciates it very much ☀️

And congratulations on the acceptance @marcsit!
